+++ pyproject.toml
@@ -12,6 +12,7 @@ dependencies = [
     "argh>=0.31.3",
     "authlib>=1.6.0",
     "bs4>=0.0.2",
+    "cryptography>=45.0.3",
     "lxml>=5.4.0",
     "nicegui>=2.17.0",
     "office365-rest-python-client>=2.6.2",
